feat(logo-grid): better support dark theme
