+++ packages/logo-grid/package.json
@@ -7,6 +7,7 @@
   "dependencies": {
     "@hashicorp/react-button": "^6.0.0",
     "@hashicorp/react-image": "^4.0.1",
+    "@hashicorp/react-inline-svg": "^6.0.3",
     "@reach/dialog": "^0.16.0",
     "@reach/portal": "^0.16.0",
     "@reach/rect": "^0.16.0",
