+++ packages/logo-grid/style.module.css
@@ -10,57 +10,28 @@
       margin: -10px;
     }
   }
+
+  /* Light theme, for logo color prop "color" or "monochrome" */
+  &.light {
+    --tile-border-color: var(--gray-5);
+    --tile-box-shadow: 0 8px 12px rgba(0, 0, 0, 0.15);
+  }
+
+  /* Dark theme, for logo color prop "white" */
+  &.dark {
+    --tile-border-color: var(--gray-2);
+    --tile-box-shadow: 0 8px 12px rgba(255, 255, 255, 0.1);
+  }
+
+  /* removeBorders overrides either theme */
+  &.removeBorders {
+    --tile-border-color: var(--transparent);
+  }
 }
 
-.gridItem {
+.listItem {
   position: relative;
   margin: 15px;
-  border: 1px solid var(--gray-5);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  & picture {
-    width: 100%;
-    height: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    & img {
-      margin: 0 auto;
-    }
-  }
-
-  & img {
-    display: block;
-    max-height: 70%;
-    max-width: 70%;
-  }
-
-  &.hasLinkOrTooltip {
-    cursor: pointer;
-    transition: 0.25s ease;
-    transition-property: border, box-shadow, transform;
-
-    &:hover {
-      border: 2px solid var(--brand);
-      box-shadow: 0 8px 12px rgba(0, 0, 0, 0.15);
-      transform: translateY(-4px);
-    }
-
-    &:focus {
-      outline: 0;
-    }
-
-    & a {
-      display: flex;
-      height: 100%;
-      width: 100%;
-      align-items: center;
-      justify-content: center;
-    }
-  }
 
   &.large {
     width: calc(33.333333% - 30px);
@@ -107,8 +78,81 @@
       margin: 10px;
     }
   }
+}
 
-  &.removeBorders {
-    border: none;
+.tilePlain {
+  width: 100%;
+  height: 100%;
+  border: 1px solid var(--tile-border-color);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background: transparent;
+
+  & picture {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    & img {
+      margin: 0 auto;
+    }
   }
+
+  & img {
+    display: block;
+    max-height: 70%;
+    max-width: 70%;
+  }
+}
+
+.tileClickable {
+  composes: tilePlain;
+  cursor: pointer;
+  transition: 0.25s ease;
+  transition-property: border, box-shadow, transform;
+
+  &:hover,
+  &.showDialog {
+    box-shadow: var(--tile-box-shadow);
+    transform: translateY(-4px);
+  }
+}
+
+/* Tooltip content */
+
+.companyName {
+  composes: g-type-display-5 from global;
+  margin: 0;
+
+  /* Light theme, for logo color prop "color" or "monochrome" */
+  &.light {
+    color: var(--black);
+  }
+
+  /* Dark theme, for logo color prop "white" */
+  &.dark {
+    color: var(--white);
+  }
+}
+
+.companyDescription {
+  composes: g-type-body-small from global;
+  margin: 1rem 0 0 0;
+
+  /* Light theme, for logo color prop "color" or "monochrome" */
+  &.light {
+    color: var(--gray-3);
+  }
+
+  /* Dark theme, for logo color prop "white" */
+  &.dark {
+    color: var(--gray-5);
+  }
+}
+
+.companyLinkButton {
+  margin-top: 1.5rem;
 }
